Fix bug in TextBuffer that causes empty inputs to not have any lines
When processing empty text, text buffer should return an instance that
contains one empty line. This ensures that callers can pass in '0' as
the lookup position. Since the lexer (and thus the parser) see text with
an additional '\0' at the end (indicating EOF) they will likely report
report error information for that location. The canonical example is:

    Found 'end-of-file' but expected 'SELECT'