Introduce SeparatedSyntaxList<TNode>
This element is not a syntax node itself. Instead, it's elements
are meant to live directly in their containing node.

This allows storing a list of syntax nodes that are separated by
a token (right now, it's a comma in all cases). This approach
replaces the previous way of storing the comma as a nullable token
on the preceding syntax node.