Ensure cascading errors are avoided
When binding expressions that depend on the type of other expressions,
such as operators, casts, function invocations, method invocations, and
property accesses, it's crucial to avoid cascading errors.

For example, when any argument of a binary operator expression can't be
resolved, it doesn't make sense to report that that the operator cannot
be applied on type <unknown>. Instead, it is better to return a bound
expression that couldn't be resolved.