Fix parser to skip bad tokens while reporting errors
The method SyntaxTree.GetDiagnostics() needs to descend into the trivia
structure in order to report diagnostics for skipped tokens.

We fix that by adding a 'descendIntoTrivia' optional parameter on the
underlying walking methods. In addition, the initial LexAllTokens method
now adds all bad tokens to the leading trivia list of the next, non
bad-token. This non-bad token is guaranteed to exist, because the input
is always terminated with a synthesized (and thus valid) EOF token.