Improve error messages for PropertyAccessExpression
When the binder cannot resolve the name of a NameExpression, it checks
whether it can find a function or aggregate and reports a 'missing
parentheses' error message instead of the 'cannot find name foo'. For
PropertyAccessExpressions the binder also tries to find a method if it
can't resolve a property.