Fix parsing of TOP x WITH TIES
A missing WITH or TIES needs to result in an error message. We also
need to make sure that either one can be missing without confusing
the parser. In those cases the parser records a missing keyword.

In addition, the parser also validates that the argument to TOP is
a valid integer.