Factor completion for keywords into a separate completion provider
Make the provider smarter by reducing the keywords to more sensible
ones. It doesn't need to be perfect. It may provide more keywords than
would be 'correct' but it should not remove valid keywords
from the list of proposals.

The only exception is for the 'AS' keyword. Disable any completion if
the current token is parsed as a user given name, such as:

    SELECT expr a|

The reason for this behavior is that it avoids accidental completions
when the alias being typed is prefixed with 'AS', such as 'Associate'.