Fix diagnostic span for unterminated tokens
The goal is to avoid incredibly long diagnostic spans if tokens aren't
properly terminated. The lexer makes the assumption that the remaining
size in the buffer is two or more. However, in certain cases there is
only one character left, for example if the input is a single quote.

Fix this by computing the minimum.