Fix tests to normalize line endings before asserting
Our test are using inline strings for before and after comparisons.

Since we're using git as the version control, the default is to commit
line-endings as LF and checkout with CR-LF. However, not all clients
use this default. In order to make the (few) tests resilient to the
behavioral differences, I've added a string normalization.