Fix binder to record parenthesized queries
In order to get the ORDER BY semantics correctly the binder needs to
skip any parenthesized queries in order to find the innermost SELECT
query. This causes the binder to not record any binding information
for the parenthesized queries to which an ORDER BY is applied. Any
requests to find a BoundQuery node for them will fail. This can be
observed by calling SemanticModel.GetOutputColumns() which doesn't
return any columns for the parenthesized queries.