Centralize trying multiple tokens in authoring features
In authoring scenarios we often need to try both, the token that
contains a given position, as well as the previous token if the
position is the end of that token. This allows correct handling
in cases like:

   x|+y

For some authoring scenarios considering 'x' makes more sense
then '+', although but isn't always correct either. For example,
the code action that swaps operands will need to see '+' token.

Long story short, we need to try them both.