Improve performance in parser
Currently, we do a bunch of checks each time Match() is called. However,
in many cases the current token is the expected token.

Also, we use Peek(-1) to get the previous token which would require
bounds checking -- which we don't even do. However, this check would
slow down a bunch of other operations that use Peek(). So instead,
introduce a new method GetPreviousToken() which can perform the check.