Add concept of a completion builder
Handle cases where we want to return a list of completion items but the
currently entered text could be a user provided name.

Returning a builder allows completion providers to include the currently
entered text as a valid entry in that list. This avoids the problem that
the other items compete with the currently entered text.

This approach is very similar to Visual Studio's suggestion mode.