Fix issues around checking if the caret is in a user provided name
We need to differentiate whether the current identifier is guaranteed to
be an alias or just possibly an alias. This is due to SQLs syntax rules
where aliases can simply be provided as a suffix, without a leading AS
keyword.

In this case

   SELECT *
   FROM   Employees
            I|

the identifier 'I' is parsed as an alias. However, typing additional
characters can change that picture:

   SELECT *
   FROM   Employees
            INNER |

In some case though we can guarantee that is an alias. For example, in
the following query 'FullName' will always be an alias:

   SELECT e.FirstName + ' ' + e.LastName AS FullName|
   FROM   Employees e