Add alias completion provider
In order to allow other providers to return completions in cases where
the currently typed identifier might be a user provided name, the
provider returns a builder for the current identifier.

So in cases like this:

    SELECT  e.LastName A|
    FROM    Employees e

we'll return a builder for 'A'. This allows the keyword provider to
also provide 'AS' without causing a commit when the user hits space. In
that case, only 'A' would be committed instead of 'AS'.