Make row buffer copying more efficient by adding an abstract member
Make copying a row buffer to an object[] a first class concept and
expose it on RowBuffer as an abstract method. This allows each row
buffer to delegate this operation to their row buffers which in turn
means we can often delegate directly to Array.Copy.

Also expose IndirectedRowBuffer as a shared internal class.