Make regular expression evaluation significantly faster
We currently evaluate all expressions as queries. That's semantically
correct because expressions may contain subqueries which requires a
query evaluation. However, many expressions don't contain subqueries.

Add the concept of an ExpressionEvaluator which is created from the
CompiledQuery object. When an expression evaluator is constructed,
check whether we have a trivial expression. If so, just compile the
first expression to a Func. Otherwise we return a Func that performs
a regular query evaluation.

Since aggregate functions also use expressions this makes aggregation
a lot faster, too.