Don't eat the enter key when completing a builder
We've two types of completion items: regular ones and builders. If we're
committing a builder, we don't want to consume the enter key -- we want
it to be send to the editor.

The reason being that the builders are designed to support aggressive
IntelliSense without interrupting the typing flow. Consuming the commit
character is violating that goal.

For example, consider the following completion scenario:

    SELECT  *
    FROM    Employees e|

When hitting enter at the caret position we want to end up here:

    SELECT  *
    FROM    Employees e
    |