Add option to suppress running tests
When running on a build machine, such as Visual Studio Online, the outer
build process already runs the tests as a secondary step. The benefit is
that those systems typically provide additional functionality, such as
measuring code coverage and reporting on the pass rate.

Thus, it makes sense to suppress running the tests as part of the
inner build process so that tests aren't executed twice.