Add Workspace and Document concept
We currently have the concept of a mutable NQueryDocument. It basically
represents the text buffer. The problem is that this can't be passed
safely to worker threads.

Instead, we should model what Roslyn did. They have a mutable Workspace
which offers an immutable snapshot.

In our case the workspace is associate with the single text buffer.
We then use the concept of a document to represent a snapshotted version
of the text as well as the editor configuration, such as the data
context and whether it's query or an expression.

The document also exposes methods to compute the syntax tree and it's
semantic model asynchronously. It also retains their results to avoid
computing it multiple times.