Make GetAppliedSelectQuery() public
Due to the semantics of ORDER BY queries it's required to be able to
get the applied SELECT query from the outside. It's better to expose
a public API rather than forcing consumers to replicate the necessary
logic.

The method is currently in the Binder class. The binder is internal
so we can't expose the method there. A more suitable way to expose
this functionality is via an extension method on OrderedQuerySyntax.