Add code fix for expressions that must be grouped
A common issue in SQL is that aggregate functions require that certain
columns must be added to GROUP BY. For example, this query is invalid:

    SELECT  e.City,
            COUNT(*)
    FROM    Employees e

This provides a code fix that will suggest adding e.City to GROUP BY:

    SELECT  e.City,
            COUNT(*)
    FROM    Employees e
    GROUP   BY e.City

Currently, we don't have the notion of a CodeFixProviders.

We only provide a refactoring provider and a code issue provider.
Refactoring providers suggest changes. Code issue providers report
issues and suggest fixes at the same time.

For simplicity, the CodeFixProvider is represented as a specialized
refactoring provider. We should probably make this a first class
concept.