Componentize outlining
Currently, the outlining is done in a singular class that handles all
cases. This isn't in line with how we process information in other
areas, such as brace matching.

Fix this by introducing the concept of an IOutliner and let specific
cases be handled in separate implementations. However, we want to
preserve the current performance profile which essentially visits the
tree only once.