Add TextChangeSet
The way we currently produce changes in from code actions isn't very
efficient. We've added extension methods on SyntaxTree which cause
each intermediary edit to be represented as full-blow syntax tree.

First of all, I've added a new method WithChanges on the SyntaxTree.
This allows the syntax tree to handle incremental parsing (if we
ever going to implement that).

I've also created a new type TextChangeSet that allows convenient
recording of edits (i.e. insertions, deletions and replacements).
The semantics of this type are simple:

* All text spans represent positions in the unedited representation.
  This frees consumers from having to translate any spans.

* It implements IEnumerable<TextChange> and will return the changes
  in reverse order so that consumer can simply apply the changes
  without having to translate any spans.

* Edits must not overlap. In practice, this shouldn't be a real
  limitation.