Simplify commenter
We no longer have to order the operations -- SourceText already takes
care of that.

However, we actually had a silent bug because we relied on the fact
that changes were performed in order. The problem is that inserting
text at the same location is ambiguous. It's easier to read if we
special case that and create a single consecutive edit.