Add tests for parsing
Instead of having a splintered set of test we should have a single set
that systematically parses the entire grammar.

In order to make validation easier, this adds an asserting enumerator
and a command to the viewer application to generate a test case from
the current query, honoring the selection if one is present.