Fix AssertingEnumerator so that Dispose doesn't mask errors
Currently, the pattern for using the AssertingEnumerator looks like
this:

    using (var enumerator = AssertingEnumerator.ForExpression(text))
    {
        enumerator.AssertNode(SyntaxKind.ExpressionSelectColumn);
        enumerator.AssertNode(SyntaxKind.AddExpression);
        enumerator.AssertNode(SyntaxKind.LiteralExpression);
        enumerator.AssertToken(SyntaxKind.NumericLiteralToken, @"1");
        enumerator.AssertToken(SyntaxKind.PlusToken, @"+");
        enumerator.AssertNode(SyntaxKind.LiteralExpression);
        enumerator.AssertToken(SyntaxKind.NumericLiteralToken, @"2");
    }

If one of those asserts throws, the enumerator will still be disposed.

Unfortunately, the Dispose method will run additional validation, i.e.
that no diagnostics are pending and the enumerator is at the end. If
those tests will fail, another exception will be thrown overriding the
current exception which makes it appear as if the Dispose method failed.

To fix this, all Assert* methods get an exception filter that will let
all exceptions through but will record the fact that an exception was
thrown, allowing the Dispose method to suppress any validation when an
error already occurred.