Improve editor application
Instead of having the lemon color command bar, all commands should reside
in the main menu. Firstly, this makes discovery of shortcuts more intutive
and secondly it's easier to extend.

Additionally, the syntax view is starting to get in the way. We don't
really need it. So instead of being permanently there, it now defaults to
be invisible unless turned on.