Consider built-in types when resolving operators
When performing overload resolution on user defined operators we need to
consider built-in operators as well. We currently assume that if we find
any user defined operators that we don't need to consider the built-in
ones at all. That's not correct. Consider a type definition like this:

    class CustomBinary
    {
        private readonly string _value;

        public CustomBinary(string value)
        {
            _value = value;
        }

        public static string operator *(CustomBinary x, CustomBinary y)
        {
            return $"{x} * {y}";
        }

        public static implicit operator int (CustomBinary arg)
        {
            return arg._value.Length;
        }
    }

Now consider an expression like this:

   left * right

If both arguments are of type CustomBinary, it will simply use the user
defined multiplication operator. But consider what happens if one of the
arguments is of type int.

In that case we'd report an error because the only operator to consider
requires both arguments to be of type CustomBinary and there is no
implicit conversion from int to CustomBinary.

That's not what C# does. C# also considers the built-in operators which
includes a multiplication operator that is defined for two ints. Since
CustomBinary has an implicit conversion to int, that's the one it would
pick.

This fixes this by making sure that overload resolution also considers
built-in operators if, and only if, no applicable user defined operator
was found.