Fix FlipBinaryOperatorSidesCodeRefactoringProvider
The provider currently simply swaps the operands and doesn't change the
operator. While there are uses for it, it's not really that helpful. In
most cases, you'd want to preserve the meaning of the expression. For
instance, swapping the operands in this case:

    e.EmployeeId > 10

should yield

    10 < e.EmployeeId

In addition to that, there are certain operators where changing sides
changes the semantics. For instance, swapping operands in these
expressions would change the output:

    left ** right
    left >> right
    left << right
    left % right
    left / right