Fix duplicated code actions
The base refactoring provider will use the FindStartTokens() method to
find suitable starts for finding the ancestors relevant for the provider.

The problem is that it can create in duplicated entries for the ancestor
node which in turn results in duplicated code actions.

This will add new FindNodes() methods that take combine calling
FindStartTokens(), followed by getting all ancestors and ensuring no
duplicated values are returned while retaining the order.