Simplify instantiation of CTEs
Instead of mapping the newly generated value slots of the CTE to match the
column instances after the CTE has been instantiated, we'll inject the
column instance value slots into the instantiator.

This allows the instantiator to simply use the value slots from the column
instances as opposed to generate new slots.