Fix bug in HashMatchIterator
The following query will construct a hash match that will use the equality
predicate for the hash match and the inequality predicate as the
remainder:

    SELECT  e1.FirstName,
            e2.FirstName
    FROM    Employees e1
                INNER JOIN Employees e2 ON e1.City = e2.City AND e1.EmployeeID <> e2.EmployeeID

Th bug is that the row buffer allocation for the remainder predicate
allocation is using the left and right input of the hash match instead of
the hash match row buffer itself. This causes issues because the row
buffer creates a virtual row buffer after reading the inputs. This causes
the predicate to refer to wrong values.