Fix ordering issue
For correctness, the expansion of full outer joins need to happen before
we remove unused value slots.

This is because we need to create fake table instance symbols. Those can't
be limited to only the columns in use. However, in order to produce the
mapping from old slots to new slots the rewriter relies on the method
BoundRelation.GetDefinedValues(). Unfortunately, those will not include
value slots that aren't actually used.