Fix incorrect caching of expressions
We replace expressions in order to get the semantics of groupings,
aggregates, and projections correctly.

The bug is that they are conceptually layers. That is, when replacing
expressions for projections, it's OK to use expressions from the groupings
and aggregates. However, it's not OK to use projections because they
aren't computed yet.