Only fade name of unused CTEs
We currently fade the entire CTE if it's not being used. That's quite
annoying when the CTE is being written -- of course it's unused. It's
perfectly fine to fade the name (it serves as the reminder that it
still needs to be consumed) but writing SQL inside an area that is
completely faded is fairly challenging.