Fix bug in join orderer
In order to build a graph between the different tables we need to index
the nodes based on the value defined value slots. We currently use the
output values, as opposed to the defined values. This is problematic
because it can result in duplicate keys in case if outer references.