Fix bug when rewriting exists queries
The issue is that we don't have a way to get access to a specific
value slot factory. Instead, we infer it from existing value slots,
which presumes that there is at least one. But in some cases, such
as EXISTS subqueries, this might not be the case, which results in
crashes.

This adds a test to catch the specific issue and also adds a TODO.
However, I'm pretty sure the issue exists in other cases as well.
This is an area we need to revisit.