Allow extending selection in separated syntax lists
When growing selections in separated syntax lists, one really wants to
first extend it to the item, then the item with its separator and then
the list itself. Unfortunately, neither item/separator pair nor the list
itself are represented as nodes which means our current algorithm isn't
able to handle that case.

This adds a provider model that allows special casing the nodes with
separated syntax lists.