Fix OuterJoinRemover so it does not optimize too aggressively
Here's a query that exhibits different behavior:

SELECT  o.OrderID
FROM    Orders o
WHERE   (
            SELECT  TOP 1 od2.ShipVia
            FROM    Orders od2
            WHERE   od2.OrderID > o.OrderID
        ) IS NOT NULL

Previously, this failed with a KeyNotFoundException when evaluated,
because the OuterJoinRemover erroneously changed the left semi join to
an inner join.

Reasoning: The OuterJoinRemover checked for any null-rejecting column
anywhere in the query, but it needs to check for only those columns
that are null-rejecting in the respective subtree, i.e. children of the
join node.